requirements.txt
aiofiles
bard
SpeechRecognition
pornhub-api==0.2.0
aiohttp==3.9.1
asyncio==3.4.3
beautifulsoup4
daxxhub
deepai
python-whois
dnspython
ffmpeg-python
gitpython
hachoir
Faker
pydub
pycountry
heroku3
apscheduler
motor==3.6.0
pillow==9.5.0
psutil
wget
pyfiglet
qrcode
pytz
future==0.18.3
search_engine_parser
py-tgcalls==0.9.7
pykeyboard
pyshorteners
pyrogram==2.0.106
python-dotenv
pyyaml==6.0.1
requests
emojis
speedtest-cli
spotipy
tgcrypto
telegraph
unidecode
yt-dlp
git+https://github.com/joetats/youtube_search@master
git+https://github.com/alexmercerind/youtube-search-python@main
gtts
git+https://github.com/AsmSafone/SafoneAPI
lexica-api==1.5.1
httpx==0.25.2
gpytranslate
openai==0.28.0
pickledb
python-telegram-bot==20.7
img2pdf
